requirements.txt
pipecat-ai[daily,openai,elevenlabs,local_smart_turn-v3]==0.0.103
python-dotenv
aiohttp
loguru
